requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-apscheduler==0.7.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.5
PyJWT==2.9.0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzlocal==5.2
